Replace try with conditional path exists
When running your code for the first time (python 3.9) the error that is produced is a read error on json.load. This error gets swallowed by your FileNotFound exception causing this code to fail on initial runs. Consider the conditional above to check to see if the file exists and if it does not create it. 

This is awesome work, stick with it. I have some additional "triggers" or tags I have mapped and will add that shortly.